requirements.txt
numpy
streamlit
pillow
opencv-contrib-python